+++ backend/arcadia/urls.py
@@ -18,13 +18,21 @@ from django.contrib import admin
 from django.urls import path, include, re_path
 from public_site.views import not_found
 import user.views as user
+from django.conf.urls.static import static
+from django.conf import settings
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path("api/user/", include("user.urls")),
+    path("api/miru/", include("miru.urls")),
     path('', include('public_site.urls')),
 
     path("login", user.loginPage),
     path("register", user.register),
+]
+
+urlpatterns += static(settings.MEDIA_URL, document_root = settings.MEDIA_ROOT)
+
+urlpatterns += [
     re_path(r'^.*$', not_found),
 ]