+++ frontend/app/components/navigation/sideNav.tsx
@@ -1,4 +1,5 @@
 'use client'
+
 import Link from "next/link";
 import { redirect, usePathname } from "next/navigation";
 import React, { useEffect, useState } from "react";